% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rw.R
\name{find_rw}
\alias{find_rw}
\title{Output the absolute pathname of the 'rw' command-line tool}
\usage{
find_rw()
}
\description{
Output the absolute pathname of the 'rw' command-line tool
}
\section{Usage from POSIX shell}{


\if{html}{\out{<div class="sourceCode sh">}}\preformatted{# Get the pathname of the 'rw' tool part of the 'rw' package
$ Rscript -e rw::find_rw

# Call the 'rw' tool
$ "$(Rscript -e rw::find_rw)" --help

# Define a Bash rw() function that locates the 'rw' tool and calls
# it forwarding all command-line arguments as-is
$ rw() \{ "$(Rscript -e rw::find_rw)" "$@"; \}
$ rw --help

# Define a Bash rw() function, where we locate the 'rw' tool once
# so that the rw() does not have to look up the location each time
$ eval "rw() \{ \\"$(Rscript -e rw::find_rw)\\" \\"\\$@\\"; \}"
$ rw --help
}\if{html}{\out{</div>}}
}

