% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giganti_data.R
\docType{data}
\name{giganti_data}
\alias{giganti_data}
\title{Giganti Data
Dataset generated using code from Giganti & Shepherd (2020)}
\format{
An object of class \code{data.frame} with 4000 rows and 8 columns.
}
\usage{
giganti_data
}
\description{
Giganti Data
Dataset generated using code from Giganti & Shepherd (2020)
}
\keyword{datasets}
