% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_data.R
\docType{data}
\name{GM.integrated}
\alias{GM.integrated}
\title{Lymphoblastoid (GM) pseudo-bulk Hi-C data}
\format{
A symmetric count matrix with dimension 860 by 860.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE117874, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
GM.integrated
}
\description{
This pseudo-bulk Hi-C data is obtained by summing up the 14 GM single-cell Hi-C count data.
}
\keyword{datasets}
