% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_data.R
\docType{data}
\name{GM6.KR}
\alias{GM6.KR}
\title{Knight-Ruiz (KR) normalized lymphoblastoid (GM) single-cell Hi-C data 6}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
GM6.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{GM6} data.
}
\keyword{datasets}
