% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_data.R
\docType{data}
\name{GM8}
\alias{GM8}
\title{Lymphoblastoid (GM) single-cell Hi-C data 8}
\format{
A symmetric count matrix with dimension 860 by 860. The proportion of observed zeros is 98.79\%.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE117874, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
GM8
}
\description{
This observed real single Hi-C data correspond to the long arm of chromosome 22 of a GM cell with 40 kb resolution (GEO accession: GSM3314368).
}
\keyword{datasets}
