% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hESC_data.R
\docType{data}
\name{hESC.bulk.KR}
\alias{hESC.bulk.KR}
\title{Knight-Ruiz (KR) normalized human embryonic stem cells bulk Hi-C data}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
hESC.bulk.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{hESC.bulk} data.
}
\keyword{datasets}
