% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hESC_data.R
\docType{data}
\name{hESC.bulk}
\alias{hESC.bulk}
\title{Human Embryonic Stem Cells (hESC) bulk Hi-C data}
\format{
A symmetric count matrix with dimension 857 by 857. The proportion of observed zeros is 73.8\%.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35156, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
hESC.bulk
}
\description{
This observed real bulk Hi-C data correspond to the long arm of chromosome 22 of the hESC dataset with 40 kb resolution (GEO accession: GSE35156).
}
\keyword{datasets}
