% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMC_data.R
\docType{data}
\name{PBMC2.KR}
\alias{PBMC2.KR}
\title{Knight-Ruiz (KR) normalized peripheral blood mononuclear cells (PBMC) single-cell Hi-C data 2}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
PBMC2.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{PBMC2} data.
}
\keyword{datasets}
