% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMC_data.R
\docType{data}
\name{PBMC8}
\alias{PBMC8}
\title{Peripheral blood mononuclear cells (PBMC) single-cell Hi-C data 8}
\format{
A symmetric count matrix with dimension 860 by 860. The proportion of observed zeros is 99.12\%.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE117874, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
PBMC8
}
\description{
This observed real single Hi-C data correspond to the long arm of chromosome 22 of a PBMC cell with 40 kb resolution (GEO accession: GSM3314383).
}
\keyword{datasets}
