% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_libs_user.R
\name{r_libs_user}
\alias{r_libs_user}
\title{Manage the R package library for webR}
\usage{
r_libs_user(path = "~/R/\%p-library/\%v")
}
\arguments{
\item{path}{(character string) The path to the R package library
on the host file system to be mounted as the R user package library
in webR. Conversion specifiers as described in \link[base:libPaths]{base::R_LIBS_USER}
are automatically expanded per \code{R.version} in webR.}
}
\value{
\code{r_libs_user()} returns the package library path for webR in
\verb{~/R/\%p-library/\%v}, where \verb{\%p} is the R platform and \verb{\%v} is the
R 'x.y' version R  in webR, e.g.
\verb{~/R/wasm32-unknown-emscripten-library/4.5}.
}
\description{
Manage the R package library for webR
}
