% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_install_packages.R
\name{rw_install_packages}
\alias{rw_install_packages}
\title{Install R packages to the R user package library path in webR}
\usage{
rw_install_packages(
  pkgs,
  lib = rw_r_libs_user(),
  repos = rw_repos(),
  contriburl = rw_contriburl(repos = repos),
  available = available.packages(contriburl = contriburl),
  dependencies = NA
)
}
\arguments{
\item{pkgs}{(character vector) Names of packages to install.}

\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, e.g., the URL of a CRAN mirror such as
    \code{"https://cloud.r-project.org"}.  For more details on
    supported URL schemes see \code{\link{url}}.

    Can be \code{NULL} to install from local files, directories or URLs:
    this will be inferred by extension from \code{pkgs} if of length one.
  }

\item{contriburl}{
    URL(s) of the contrib sections of the repositories.  Use this
    argument if your repository mirror is incomplete, e.g., because
    you mirrored only the \file{contrib} section, or only have
    binary packages.  Overrides argument \code{repos}.
    Incompatible with \code{type = "both"}.
  }

\item{available}{
    a matrix as returned by \code{\link[utils]{available.packages}}
    listing packages available at the repositories, or \code{NULL} when
    the function makes an internal call to \code{available.packages}.
    Incompatible with \code{type = "both"}.
  }

\item{dependencies}{logical indicating whether to also install
    uninstalled packages which these packages depend on/link
    to/import/suggest (and so on recursively).
    Not used if \code{repos = NULL}.
    Can also be a character vector, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.

    Only supported if \code{lib} is of length one (or missing),
    so it is unambiguous where to install the dependent packages.  If
    this is not the case it is ignored, with a warning.

    The default, \code{NA}, means
    \code{c("Depends", "Imports", "LinkingTo")}.

    \code{TRUE} means to use
    \code{c("Depends", "Imports", "LinkingTo", "Suggests")} for
    \code{pkgs} and
    \code{c("Depends", "Imports", "LinkingTo")} for added dependencies:
    this installs all the packages needed to run \code{pkgs}, their
    examples, tests and vignettes (if the package author specified them
    correctly).

    In all of these, \code{"LinkingTo"} is omitted for binary packages.
  }
}
\value{
A named logical vector, where the names correspond to packages installed
and the values whether they installed successfully.
}
\description{
Install R packages to the R user package library path in webR
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Install the 'future' package
rw_install_packages("future")

# List all installed packages
pkgs <- rw_do_call(utils::installed.packages, r_libs = rw_r_libs_user())
print(rownames(pkgs))
\dontshow{\}) # examplesIf}
}
