% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm_eval_future.R
\name{rwasm_eval_future}
\alias{rwasm_eval_future}
\title{Evaluates a future in R WebAssembly (Wasm)}
\usage{
rwasm_eval_future(future, libs = NULL, shared = NULL, debug = FALSE)
}
\arguments{
\item{future}{A \link[future:Future-class]{future::Future} object.}

\item{libs}{(character string; optional) The path to an R user library
of Emscripten packages binary mounted as read-write in Rwasm.}

\item{shared}{(character string; optional) The path to a directory
mounted read-write as \verb{./shared/} to Rwasm.}

\item{debug}{(logical) If TRUE, RWasm outputs debug messages.}
}
\value{
A \link[future:FutureResult]{future::FutureResult} object.
}
\description{
Evaluates a future in R WebAssembly (Wasm)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(future)
f <- future({ sessionInfo() }, lazy = TRUE)
r <- rwasm_eval_future(f)
\dontshow{\}) # examplesIf}
}
