% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm_source.R
\name{rwasm_source}
\alias{rwasm_source}
\title{Source R script in R WebAssembly (Wasm)}
\usage{
rwasm_source(
  main,
  prologue = NULL,
  epilogue = NULL,
  shared = NULL,
  binds = character(0L),
  libs = r_libs_user(),
  debug = FALSE
)
}
\arguments{
\item{main, prologue, epilogue}{(character string)
The main R script, and optional R scripts that are sourced before and
after to the main Rscript.
If \code{AsIs} string vectors, then they are interpreted as R code, which is
then written to temporary R scripts, and sourced that way.}

\item{shared}{(character string; optional) A folder shared in webR with
pro- and epilogue scripts as \verb{/host/shared/}. The folder is \emph{not} mounted
when the main R script is sourced.}

\item{binds}{(character vector) A named vector of directories on the
host system, where the names are the paths to be mounted in webR.
These folders are available throughout the full call, i.e. during the
prologue, the main, and the epilogue steps.
Due to limitations in WebAssembly, it is \emph{not} possible to mount
directories as being read-only.}

\item{libs}{(character string; optional) The path to an R user library
of Emscripten packages binary mounted as read-write in Rwasm.}

\item{debug}{(logical) If TRUE, RWasm outputs debug messages.}
}
\value{
The captured output (standard output and standard error) as a
character string.
}
\description{
Source R script in R WebAssembly (Wasm)
}
\section{Sandboxed evaluation of the main R code}{

Note that webR, and therefore the main, the prologue and the epilogue R
code, has access to the internet. Because webR runs within WebAssembly,
such access is limited to certain protocols. See the webR and WebAssembly
documentations for further details.

The R instance running in webR does not have access to the hosts file
system, unless some directories are explicitly shared via the \code{binds}
argument.

The prologue and epilogue R codes that are run in webR before and after
the main R code do have access, while evaluated, to host directory
\code{shared}, if specified. The main R script does \emph{not} have to this
directory. Because of this, you can use \emph{trusted} prologue and epilogue
code to share data files between R on the host and webR, without giving
the main R code, which you might not trust, access.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Prove that we are running within R Wasm
code <- 'sessionInfo()'
out <- rwasm_source(I(code))
writeLines(out)

# Bind host's temporary folder
code <- 'cat("hello from webR\n", file = "/host/tmp/hello.txt")'
out <- rwasm_source(I(code), binds = c("/host/tmp" = tempdir()))
msg <- readLines(file.path(tempdir(), "hello.txt"))
writeLines(msg)

# Pass shared data files via pro- and epilogue scripts
saveRDS(list(a = 1, b = 2), file.path(tempdir(), "in.rds"))
prologue <- 'data_in <- readRDS("/host/shared/in.rds")'
epilogue <- 'saveRDS(data_out, "/host/shared/out.rds")'
code <- c(
  'data_out <- lengths(data_in)',
  'stopifnot(length(dir("/host/shared")) == 0)' ## empty => not mounted
)
output <- rwasm_source(I(code), prologue = I(prologue), epilogue = I(epilogue), shared = tempdir())
out <- readRDS(file.path(tempdir(), "out.rds"))
print(out)

# Note that R Wasm has internet access by default
code <- 'writeLines(readLines("https://ipinfo.io/json", warn = FALSE))'
out <- rwasm_source(I(code))
writeLines(out)
\dontshow{\}) # examplesIf}
}
