% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1_data.R
\docType{data}
\name{sim1.P.KR}
\alias{sim1.P.KR}
\title{Knight-Ruiz (KR) normalized \eqn{P} data from Simulation Study 1}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim1.P.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{sim1.P.obs} data.
}
\keyword{datasets}
