% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1_data.R
\docType{data}
\name{sim1.P.obs}
\alias{sim1.P.obs}
\title{Simulated \code{P} data from Simulation Study 1}
\format{
A symmetric square matrix with a clear 5-diagonal-block patterns and random background noise.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim1.P.obs
}
\description{
\code{P} is a Hi-C data with \code{N=200} bins and \code{k=5} TADs. The domain sizes are specified as 50, 30, 20, 90, and 10, respectively.
}
\keyword{datasets}
