% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2_data.R
\docType{data}
\name{sim2.E.KR}
\alias{sim2.E.KR}
\title{Knight-Ruiz (KR) normalized \code{E} data from Simulation Study 2}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim2.E.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{sim2.E} data.
}
\keyword{datasets}
