% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2_data.R
\docType{data}
\name{sim2.true.bound.E}
\alias{sim2.true.bound.E}
\title{Domain boundary points of the \eqn{E} data from Simulation Study 2}
\format{
An integer vector marking the endpoints of the 20 TADs.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim2.true.bound.E
}
\description{
The \code{N=1000} bins in the \code{sim2.E} data are divided consecutively into \code{k=20} TADs. The 20 endpoints are listed here.
}
\keyword{datasets}
