% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim3_data.R
\docType{data}
\name{sim3.bulk.KR}
\alias{sim3.bulk.KR}
\title{Knight-Ruiz (KR) normalized bulk data from Simulation Study 3}
\format{
A symmetric square matrix, with row sums (and column sums) all equal to 1.
}
\source{
https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim3.bulk.KR
}
\description{
This data is obtained by performing KR-normalization on the \code{sim3.bulk} data.
}
\keyword{datasets}
