% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim3_data.R
\docType{data}
\name{sim3.bulk}
\alias{sim3.bulk}
\title{Bulk data from Simulation Study 3}
\format{
A symmetric count matrix with dimension 220 by 220. The proportion of observed zeros is 79.2\%.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM2109887, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim3.bulk
}
\description{
This observed real bulk Hi-C data correspond to the long arm of chromosome 16 of the K562 bulk A dataset with 200 kb resolution (GEO accession: GSM2109887).
}
\keyword{datasets}
