% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim3_data.R
\docType{data}
\name{sim3.sub}
\alias{sim3.sub}
\title{One subsampled data from Simulation Study 3}
\format{
A symmetric count matrix with dimension 220 by 220. The proportion of observed zeros is 95.5\%, which is about the same as a high-quality single-cell Hi-C matrix.
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM2109887, https://github.com/osu-stat-gen/RWS-RWR
}
\usage{
sim3.sub
}
\description{
This data is obtained by subsampling 10\% of the observed counts in \code{sim3.bulk}.
}
\keyword{datasets}
