% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_r_info.R, R/rw_version.R
\name{rw_r_info}
\alias{rw_r_info}
\alias{version-functions}
\alias{rw_version}
\alias{rw_webr_version}
\alias{rw_r_version}
\title{Get R version information on the R instance that webR provides}
\usage{
rw_r_info(template = NULL)

rw_version()

rw_webr_version()

rw_r_version()
}
\arguments{
\item{template}{(optional; character string) A template with conversion
specifiers as described in \link[base:libPaths]{base::R_LIBS_USER} that are expanded.}
}
\value{
\code{rw_r_info()} returns \code{R.Version()} in webR and \code{rw_r_info(template)}
returns a character string based on the format in webR.

\code{rw_version()} returns the version of the \code{rw} CLI tool as a
\code{numeric_version} object.

\code{rw_webr_version()} returns the version of webR as a
\code{numeric_version} object.

\code{rw_r_version()} returns the R version running in webR as a
\code{numeric_version} object.
}
\description{
Functions to retrieve version information about the \code{rw} CLI, webR, and R.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
rw_r_info()
rw_r_info("~/R/\%p-library/\%v")
\dontshow{\}) # examplesIf}
\dontrun{
rw_version()
rw_webr_version()
rw_r_version()
}

}
\keyword{internal}
