% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_do_call.R
\name{rw_do_call}
\alias{rw_do_call}
\title{Call an R Function in R WASM}
\usage{
rw_do_call(
  fcn,
  args = list(),
  r_libs = Sys.getenv("RW_R_LIBS_USER", NA_character_)
)
}
\arguments{
\item{fcn}{A function.}

\item{args}{A list of arguments passed to \code{fcn}.}

\item{r_libs}{(optional) The R package library used by \code{rw}.}
}
\value{
The object from call \code{do.call(fcn, args = args)} in R WASM.
}
\description{
Call an R Function in R WASM
}
\examples{
# Evaluate Sys.info() in R WASM
info <- rw_do_call(Sys.info)
print(info)

}
