% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_r_libs_user.R
\name{rw_r_libs_user}
\alias{rw_r_libs_user}
\title{Get the R Package Library Path for webR}
\usage{
rw_r_libs_user(path = "~/R/\%p-library/\%v")
}
\arguments{
\item{path}{(character string) A path template with optional conversion
specifiers \verb{\%p} (platform) and \verb{\%v} (R version). These are expanded
based on the webR R version.}
}
\value{
A character string with the expanded library path, e.g.
\verb{~/R/wasm32-unknown-emscripten-library/4.5}.
}
\description{
Returns the conventional path for the R user package library for webR
on the host file system.
}
\section{Environment Variables}{

If the \code{RW_R_LIBS_USER} environment variable is set, it is returned
instead of the default path.
}

