% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm_eval_future.R
\name{rwasm_eval_future}
\alias{rwasm_eval_future}
\title{Evaluates a future in R WebAssembly (Wasm)}
\usage{
rwasm_eval_future(future, libs = NULL, shared = tempdir())
}
\arguments{
\item{future}{A \link[future:Future-class]{future::Future} object.}
}
\value{
A \link[future:FutureResult]{future::FutureResult} object.
}
\description{
Evaluates a future in R WebAssembly (Wasm)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(future)
f <- future({ sessionInfo() }, lazy = TRUE)
r <- rwasm_eval_future(f)
\dontshow{\}) # examplesIf}
}
