% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm_rscript.R
\name{rwasm_rscript}
\alias{rwasm_rscript}
\title{Execute code in R WebAssembly (Wasm)}
\usage{
rwasm_rscript(file = NULL, code = NULL, libs = NULL)
}
\arguments{
\item{file}{(character string) An R script.}

\item{code}{(character vector) R code.}

\item{libs}{(character string; optional) The path to an R user library
of Emscripten packages binary.}
}
\value{
The captured output (standard output and standard error) as a
character string.
}
\description{
Execute code in R WebAssembly (Wasm)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Prove that we are running within R Wasm
out <- rwasm_rscript(code = 'sessionInfo()')
writeLines(out)

# Note that R Wasm has internet access by default
out <- rwasm_rscript(code = 'writeLines(readLines("https://ipinfo.io/json", warn = FALSE))')
writeLines(out)
\dontshow{\}) # examplesIf}
}
