% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{rwasm_rscript}
\alias{rwasm_rscript}
\title{Execute code in R WebAssembly (WASM)}
\usage{
rwasm_rscript(file = NULL, code = NULL)
}
\arguments{
\item{file}{(character string) An R script.}

\item{code}{(character vector) R code.}
}
\value{
The captured output (standard output and standard error).
}
\description{
Execute code in R WebAssembly (WASM)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
out <- rwasm_rscript(code = "sessionInfo()")
writeLines(out)
\dontshow{\}) # examplesIf}
}
