% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{rwasm_rscript}
\alias{rwasm_rscript}
\title{Execute code in R WebAssembly (WASM)}
\usage{
rwasm_rscript(code)
}
\arguments{
\item{code}{(character vector) R code to be evaluated in R WASM.}
}
\value{
The captured output (standard output and standard error).
}
\description{
Execute code in R WebAssembly (WASM)
}
