% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm.R
\name{webr_version}
\alias{webr_version}
\alias{r_info}
\title{Get the webR version}
\usage{
webr_version()

r_info(template = NULL)
}
\arguments{
\item{template}{(optional; character string) A template with conversion
specifiers as described in \link[base:libPaths]{base::R_LIBS_USER} that are expanded.}
}
\value{
\code{webr_version()} returns the webR version.

\code{r_version()} returns \code{R.version()} in webR and \code{r_version(template)}
returns a character string based on the format in webR.
}
\description{
Get the webR version

Get R version information on the R instance that webR provides
}
\details{
\code{r_version()} takes 2-3 seconds the first time it is called. This
is because it queries \code{getRversion()} in webR.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
r_version()
r_version("x.y")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
r_info()
r_info("~/R/\%p-library/\%v")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
