% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm.R
\name{webr_version}
\alias{webr_version}
\alias{r_version}
\title{Get the webR version}
\usage{
webr_version()

r_version(format = NULL)
}
\arguments{
\item{format}{(optional; character string) If \code{"x.y"}, then the
'x.y' component of the R version 'x.y.z' is returned.}
}
\value{
\code{webr_version()} returns the webR version.

\code{r_version()} returns the version of R that webR implements.
}
\description{
Get the webR version

Get the R version that webR provides
}
\details{
\code{r_version()} takes 2-3 seconds the first time it is called. This
is because it queries \code{getRversion()} in webR.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
webr_version()
r_version()
r_version("x.y")
\dontshow{\}) # examplesIf}
}
